Move environment trait to separate module
